requirements.txt
alembic==0.9.6
Babel==2.5.1
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
dill==0.3.3
dominate==2.3.1
Flask==1.0.2
Flask-Babel==0.11.2
Flask-Bootstrap==3.3.7.1
Flask-KVSession==0.6.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
guess-language-spirit==0.5.3
gurobipy==9.0.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.2.0
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.3.2
multiprocess==0.70.11.1
nose==1.3.7
numpy==1.19.2
pandas==1.1.3
Pillow==8.0.0
ply==3.11
PyJWT==1.5.3
Pyomo==5.7.1
pyparsing==2.4.7
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
pytz==2017.2
PyUtilib==6.0.0
redis==3.5.3
requests==2.18.4
scipy==1.5.2
seaborn==0.11.0
simplekv==0.14.1
six==1.11.0
SQLAlchemy==1.1.14
sqlalchemy-json==0.4.0
tabulate==0.8.8
urllib3==1.22
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
xlrd==2.0.1
